Fix error on server start when config doesn't exist yet

+++ util/fcm.js
@@ -64,12 +64,12 @@ module.exports = {
     * */
     initialize: async (strapi) => {
         // console.log('initialize FCM');
-        const { serviceAccount } = await strapi.db.query('plugin::strapi-plugin-fcm.fcm-plugin-configuration').findOne({
+        const data = await strapi.db.query('plugin::strapi-plugin-fcm.fcm-plugin-configuration').findOne({
             select: ['serviceAccount']
         });
         // console.log('serviceAccount', serviceAccount);
         // console.log('admin.apps?.length', admin.apps?.length);
-        if (serviceAccount) {
+        if (data !== null && data.serviceAccount) {
             if (admin.apps?.length > 1) {
                 Promise.all(admin.apps.map(app => app.delete())).then(() => {
                     admin.initializeApp({
